worker/requirements.txt
redis==7.0.1
httpx==0.28.1
python-dotenv==1.2.1
fastmcp>=2.13.0.2
